+++ cli/srecli/group/sre.py
@@ -1,5 +1,5 @@
 import click
-
+from srecli import __version__
 from srecli.group import ContextObj, ViewCommand, cli_response, response
 from srecli.group.customer import customer
 from srecli.group.integrations import integrations
@@ -10,7 +10,7 @@ from srecli.group.metrics import metrics
 from srecli.group.platform import platform
 from srecli.group.policy import policy
 from srecli.group.report import report
-from srecli.group.results import results
+from srecli.group.resource import resource
 from srecli.group.role import role
 from srecli.group.rule import rule
 from srecli.group.ruleset import ruleset
@@ -19,11 +19,11 @@ from srecli.group.service import service
 from srecli.group.setting import setting
 from srecli.group.tenant import tenant
 from srecli.group.users import users
-from srecli.group.resource import resource
-
-from srecli.service.helpers import validate_api_link, check_version_compatibility
+from srecli.service.helpers import (
+    check_version_compatibility,
+    validate_api_link,
+)
 from srecli.service.logger import get_logger
-from srecli import __version__
 
 
 _LOG = get_logger(__name__)
@@ -144,7 +144,6 @@ sre.add_command(rulesource)
 sre.add_command(license)
 sre.add_command(service)
 sre.add_command(setting)
-sre.add_command(results)
 sre.add_command(metadata)
 sre.add_command(metrics)
 sre.add_command(platform)