+++ src/lambdas/api_handler/handler.py
@@ -5,29 +5,26 @@ from typing_extensions import Self
 from handlers import Mapping
 from helpers import RequestContext
 from helpers.log_helper import get_logger
-from lambdas.api_handler.handlers.batch_result_handler import (
-    BatchResultsHandler,
-)
 from lambdas.api_handler.handlers.events_handler import EventsHandler
 from lambdas.api_handler.handlers.health_check_handler import (
     HealthCheckHandler,
 )
 from lambdas.api_handler.handlers.job_handler import JobHandler
+from lambdas.api_handler.handlers.new_swagger_handler import SwaggerHandler
 from lambdas.api_handler.handlers.service_operations_status_handler import (
     ServiceOperationsStatusHandler,
 )
-from lambdas.api_handler.handlers.new_swagger_handler import \
-    SwaggerHandler
 from lambdas.api_handler.handlers.users_handler import UsersHandler
 from services.abs_lambda import (
+    ApiEventProcessorLambdaHandler,
     ApiGatewayEventProcessor,
     CheckPermissionEventProcessor,
-    RestrictCustomerEventProcessor,
     ExpandEnvironmentEventProcessor,
-    ApiEventProcessorLambdaHandler,
-    RestrictTenantEventProcessor
+    RestrictCustomerEventProcessor,
+    RestrictTenantEventProcessor,
 )
 from validators.registry import permissions_mapping
+
 
 _LOG = get_logger(__name__)
 
@@ -43,7 +40,6 @@ class ApiHandler(ApiEventProcessorLambdaHandler):
 
     def __init__(self):
         self.additional_handlers = [
-            BatchResultsHandler,
             UsersHandler,
             JobHandler,
             HealthCheckHandler,