+++ src/services/job_service.py
@@ -1,16 +1,16 @@
 import uuid
-from datetime import datetime, timedelta
+from datetime import date, datetime, timedelta
 from typing import Any, Callable
 
 from pynamodb.expressions.condition import Condition
 from pynamodb.pagination import ResultIterator
 
-from helpers.constants import JobState
+from helpers.constants import JobState, JobType
 from helpers.time_helper import utc_iso
 from models.job import Job
 from services.base_data_service import BaseDataService
-from services.ruleset_service import RulesetName
 from services.platform_service import Platform
+from services.ruleset_service import RulesetName
 
 
 class JobService(BaseDataService[Job]):
@@ -25,6 +25,7 @@ class JobService(BaseDataService[Job]):
         ttl: timedelta | None = None,
         owner: str | None = None,
         affected_license: str | None = None,
+        job_type: JobType = JobType.STANDARD,
         status: JobState = JobState.SUBMITTED,
         batch_job_id: str | None = None,
         celery_job_id: str | None = None,
@@ -44,6 +45,7 @@ class JobService(BaseDataService[Job]):
             ttl=ttl,
             owner=owner,
             affected_license=affected_license,
+            job_type=job_type.value,
             status=status.value,
             batch_job_id=batch_job_id,
             celery_job_id=celery_job_id,
@@ -101,6 +103,8 @@ class JobService(BaseDataService[Job]):
     def get_by_customer_name(
         self,
         customer_name: str,
+        job_id: str | None = None,
+        job_type: JobType | None = None,
         status: JobState | None = None,
         start: datetime | None = None,
         end: datetime | None = None,
@@ -120,8 +124,50 @@ class JobService(BaseDataService[Job]):
             rkc = None
         if status:
             filter_condition &= Job.status == status.value
+        if job_id:
+            filter_condition &= Job.id == job_id
+        if job_type:
+            filter_condition &= Job.job_type == job_type.value
         return Job.customer_name_submitted_at_index.query(
             hash_key=customer_name,
+            range_key_condition=rkc,
+            filter_condition=filter_condition,
+            scan_index_forward=ascending,
+            limit=limit,
+            last_evaluated_key=last_evaluated_key,
+        )
+
+    def get_by_job_type(
+        self,
+        job_type: JobType,
+        job_id: str | None = None,
+        customer_name: str | None = None,
+        status: JobState | None = None,
+        start: datetime | date | None = None,
+        end: datetime | date | None = None,
+        filter_condition: Condition | None = None,
+        ascending: bool = False,
+        limit: int | None = None,
+        last_evaluated_key: dict | None = None,
+    ) -> ResultIterator[Job]:
+        if start and end:
+            rkc = Job.submitted_at.between(utc_iso(start), utc_iso(end))
+        elif start:
+            rkc = Job.submitted_at >= utc_iso(start)
+        elif end:
+            rkc = Job.submitted_at <= utc_iso(end)
+        else:
+            rkc = None
+        if status:
+            filter_condition &= Job.status == status.value
+        if job_id:
+            filter_condition &= Job.id == job_id
+        if customer_name:
+            filter_condition &= Job.customer_name == customer_name
+        if job_type:
+            filter_condition &= Job.job_type == job_type.value
+        return Job.query(
+            hash_key=job_type,
             range_key_condition=rkc,
             filter_condition=filter_condition,
             scan_index_forward=ascending,
@@ -132,9 +178,10 @@ class JobService(BaseDataService[Job]):
     def get_by_tenant_name(
         self,
         tenant_name: str,
+        job_type: JobType | None = None,
         status: JobState | None = None,
-        start: datetime | None = None,
-        end: datetime | None = None,
+        start: datetime | date | None = None,
+        end: datetime | date | None = None,
         filter_condition: Condition | None = None,
         ascending: bool = False,
         limit: int | None = None,
@@ -150,6 +197,8 @@ class JobService(BaseDataService[Job]):
             rkc = None
         if status:
             filter_condition &= Job.status == status.value
+        if job_type:
+            filter_condition &= Job.job_type == job_type.value
         return Job.tenant_name_submitted_at_index.query(
             hash_key=tenant_name,
             range_key_condition=rkc,